Apt: the const-ctx chain + dword count widths -- first live parse SUCCEEDS
The activation's first boot hard-crashed in LoadAnimation. Two root causes,
both found and fixed against the Xbox One assembly:

1. DWORD count reads. The XB1 parse reads every record count as a signed
   DWORD (case 142/155: `cmp [r8+8], r9d`; case 136/150: `cmp [rsi], edi`;
   const-record type: `mov eax,[rcx]`). The i64 reads of the DefineFunction/2
   nArgs slot dragged the u16 register-count/preload-flags pack (+0x0C/+0x0E)
   into the high dword -> giant arg loop -> AV. All count/type/payload reads
   are now the asm-faithful i32/dword forms (incl. the vucomiss NaN ->
   Integer(0) float quirk), and the ByteCode structs carry u32+pad instead
   of i64 for the count/body-length fields.

2. The parse ctx is the "Apt constant file" CHUNK, not the aptdata root.
   Decompiled the XB1 chain end-to-end: CompleteLoad (sub_1408348B0) reads
   movieOffset @const+0x18 (+pBase = the movie root -- retires the
   LocateMovieRoot8 signature scan on the primary path), relocates
   itemStart @const+0x28 (the 16-byte {u32 type, pad, payload} record table
   _parseStream walks; string payloads rebase ctx-relative), zeroes the def's
   +0x50 parsed-value accumulator (mpUnk50 -> mnParsedValueCount), then
   Fixup(def, base, CONST, a5) -> the resolve twin (sub_1408567D0) threads
   (ctx, def+0x50) into every _parseStream call. libapt2's Const::Write
   layout matches the XB1 const chunk EXACTLY (verified writer-side).
   Un-collapsed pBase/pConstFile at both host sites (the libapt2 6-field
   header: aptData@+0x10, const@+0x18 -- the old +16 read was the aptdata
   chunk); Resolve now relocates ctx+0x28 around Fixup (the console's
   `*(a3+28) +=/-= a3` widened); resolve64 gains the (pConstFile,
   pnParsedValues) params.

Also: BrnGameModule's PC stand-in IO-buffer stacks sized to the DOCUMENTED
X360 values (0x780000 each, main() @0x827E60D8) -- the 64KB scratch blocks
overflowed once the sound module's ~66KB Root IO buffers landed (the
stage-4 boot blocker).

Boot-verified: 0 asserts, 61 action streams parse (title + 5 imports, all
opcodes within the mapped set), Fixup completes (41 chars / 5 imports),
title_screen02 ticks to frame 100 with a 12-child display list.

+++ src/SDKs/EATech/include/Apt/AptActionInterpreterStackOps.cpp
@@ -231,16 +231,17 @@ void AptActionInterpreter::_FunctionAptActionPushFloat(AptActionInterpreter* pIn
 void AptActionInterpreter::_FunctionAptActionPush(AptActionInterpreter* pInterp, LocalContextT* pCtx)
 {
     // The GUIAPT64 operand block (the form _parseStream produces): 8-aligned
-    // {i64 count @0, AptValue** table @+8}, PC advances 16. The table entries are
-    // LIVE AptValue*s post-parse (console form was 4-aligned {i32, ptr} +8).
+    // {i32 count @0 (the xb1 parse reads it as a signed DWORD -- `cmp [rsi], edi`),
+    // AptValue** table @+8}, PC advances 16. The table entries are LIVE AptValue*s
+    // post-parse (console form was 4-aligned {i32, ptr} +8).
     const unsigned char* pAligned =
         reinterpret_cast<const unsigned char*>(
             (reinterpret_cast<uintptr_t>(pCtx->mpProgramCounter) + 7) & ~static_cast<uintptr_t>(7));
-    const int64_t   nCount = *reinterpret_cast<const int64_t*>(pAligned);
+    const int32_t   nCount = *reinterpret_cast<const int32_t*>(pAligned);
     AptValue* const* pArray = *reinterpret_cast<AptValue* const* const*>(pAligned + 8);
     pCtx->mpProgramCounter = pAligned + 16;
 
-    for (int64_t i = 0; i < nCount; ++i)
+    for (int32_t i = 0; i < nCount; ++i)
         pInterp->stackPushIndirect(pArray[i]);   // real member (resolves Lookup/Register)
 }
